+++ Minecraft.Client/Stitcher.cpp
@@ -122,7 +122,7 @@ int Stitcher::smallestEncompassingPowerOfTwo(int input)
 
 bool Stitcher::addToStorage(TextureHolder *textureHolder)
 {
-	for (int i = 0; i < storage.size(); i++)
+	for (size_t i = 0; i < storage.size(); i++)
 	{
 		if (storage.at(i)->add(textureHolder))
 		{
